feat: update getAllTags
