+++ src/posts/models/tags.model.ts
@@ -1,0 +1,8 @@
+import { Field, ObjectType } from '@nestjs/graphql'
+import { PostItemModel } from './post.model'
+
+@ObjectType()
+export class TagsModel {
+  @Field(() => [String])
+  public readonly tags: string[]
+}
